Audio does get toggled now.
