+++ src/components/index.js
@@ -4,5 +4,6 @@ import ToolBox from './ToolBox'
 import AnswerPanel from './AnswerPanel'
 import ActionBar from './ActionBar'
 import QuizPicker from './QuizPicker'
+import QuestionReader from './QuestionReader'
 
-export { QuestionBar, ToolBox, QuizzerPopup, AnswerPanel, ActionBar, QuizPicker }
+export { QuestionBar, ToolBox, QuizzerPopup, AnswerPanel, ActionBar, QuizPicker, QuestionReader }
